display pie chart functionality working successfully

+++ src/app/token.service.ts
@@ -1,21 +1,43 @@
 import { Injectable } from '@angular/core';
+import jwt_decode from "jwt-decode";
+
+
+
+
 
 @Injectable({
   providedIn: 'root'
 })
 export class TokenStorageService {
-  constructor() { }
-  signOut(): void {
-    localStorage.clear();
-  }
-   saveToken(token: string): void {
 
+  isClicked!:boolean;
+  
+  constructor() { 
+   
+  }
+  signOut() {
+    if(  localStorage.getItem('token') !=null){
+    localStorage.clear();
+    }else sessionStorage.clear();
+   
+  }
+   saveToken(token: string,isClick:boolean): void {
+      this.isClicked=isClick;
+      //console.log("in save token "+isClick)
+    if(isClick){
     localStorage.removeItem(token);
     localStorage.setItem('token', token);
+    }else{
+      sessionStorage.setItem('token1',token);
+    }
     
   }
    getToken(): string | null {
-    return localStorage.getItem('token');
+     //console.log("in get token "+this.isClicked)
+     if( localStorage.getItem('token') !=null){
+     return  localStorage.getItem('token');
+     }else return sessionStorage.getItem('token1');
+     
   }
    saveUser(user: any) {
     localStorage.removeItem('user');
@@ -28,6 +50,16 @@ export class TokenStorageService {
     }
     return {};
   }
+
+
+  decodeToken():any{
+    // if(this.getToken !==null){
+      const decoded = jwt_decode(this.getToken() as string);
+      console.log(decoded)
+      return decoded;
+    // }
+  }
+
